+++ apps/public-docsite/src/pages/Controls/DetailsListPage/DetailsListCustomRowsPage.doc.ts
@@ -1,12 +1,9 @@
 import { TFabricPlatformPageProps } from '../../../interfaces/Platforms';
 import { DetailsListCustomRowsPageProps as ExternalProps } from '@fluentui/react-examples/lib/react/DetailsList/DetailsList.doc';
 
-const related = require('!raw-loader?esModule=false!@fluentui/public-docsite/src/pages/Controls/DetailsListPage/docs/DetailsListRelated.md') as string;
-
 export const DetailsListCustomRowsPageProps: TFabricPlatformPageProps = {
   web: {
     ...(ExternalProps as any),
     title: 'DetailsList - Custom Item Rows',
-    related,
   },
 };