+++ packages/make-styles/src/runtime/resolveStyleRules.test.ts
@@ -20,7 +20,7 @@ expect.addSnapshotSerializer({
 });
 
 function getFirstClassName(resolvedStyles: Record<string, MakeStylesResolvedRule>): string {
-  return resolvedStyles[Object.keys(resolvedStyles)[0]][0];
+  return resolvedStyles[Object.keys(resolvedStyles)[0]][0] as string;
 }
 
 describe('resolveStyleRules', () => {
@@ -367,8 +367,28 @@ describe('resolveStyleRules', () => {
     });
   });
 
-  describe('experimental', () => {
-    it('allows to define keyframes', () => {
+  describe('keyframes', () => {
+    it('allows to define string as animationName', () => {
+      expect(
+        resolveStyleRules({
+          animationName: 'fade-in slide-out',
+          animationIterationCount: 'infinite',
+          animationDuration: '5s',
+        }),
+      ).toMatchInlineSnapshot(`
+        .fc59ano0 {
+          animation-name: fade-in slide-out;
+        }
+        .f1cpbl36 {
+          animation-iteration-count: infinite;
+        }
+        .f1t9cprh {
+          animation-duration: 5s;
+        }
+      `);
+    });
+
+    it('allows to define object as animationName', () => {
       expect(
         resolveStyleRules({
           animationName: {
@@ -383,26 +403,191 @@ describe('resolveStyleRules', () => {
           animationDuration: '5s',
         }),
       ).toMatchInlineSnapshot(`
-        @keyframes f13owpa8 {
-          from {
-            transform: rotate(0deg);
-          }
-          to {
-            transform: rotate(360deg);
-          }
-        }
-        .fkf6eed0 {
-          animation-name: f13owpa8;
-        }
-        .f1cpbl36 {
-          animation-iteration-count: infinite;
-        }
-        .f1t9cprh {
-          animation-duration: 5s;
-        }
-      `);
+              @-webkit-keyframes f13owpa8 {
+                from {
+                  -webkit-transform: rotate(0deg);
+                  -moz-transform: rotate(0deg);
+                  -ms-transform: rotate(0deg);
+                  transform: rotate(0deg);
+                }
+                to {
+                  -webkit-transform: rotate(360deg);
+                  -moz-transform: rotate(360deg);
+                  -ms-transform: rotate(360deg);
+                  transform: rotate(360deg);
+                }
+              }
+              @keyframes f13owpa8 {
+                from {
+                  -webkit-transform: rotate(0deg);
+                  -moz-transform: rotate(0deg);
+                  -ms-transform: rotate(0deg);
+                  transform: rotate(0deg);
+                }
+                to {
+                  -webkit-transform: rotate(360deg);
+                  -moz-transform: rotate(360deg);
+                  -ms-transform: rotate(360deg);
+                  transform: rotate(360deg);
+                }
+              }
+              @-webkit-keyframes rf13owpa8 {
+                from {
+                  -webkit-transform: rotate(0deg);
+                  -moz-transform: rotate(0deg);
+                  -ms-transform: rotate(0deg);
+                  transform: rotate(0deg);
+                }
+                to {
+                  -webkit-transform: rotate(-360deg);
+                  -moz-transform: rotate(-360deg);
+                  -ms-transform: rotate(-360deg);
+                  transform: rotate(-360deg);
+                }
+              }
+              @keyframes rf13owpa8 {
+                from {
+                  -webkit-transform: rotate(0deg);
+                  -moz-transform: rotate(0deg);
+                  -ms-transform: rotate(0deg);
+                  transform: rotate(0deg);
+                }
+                to {
+                  -webkit-transform: rotate(-360deg);
+                  -moz-transform: rotate(-360deg);
+                  -ms-transform: rotate(-360deg);
+                  transform: rotate(-360deg);
+                }
+              }
+              .fkf6eed0 {
+                animation-name: f13owpa8;
+              }
+              .rfkf6eed0 {
+                animation-name: rf13owpa8;
+              }
+              .f1cpbl36 {
+                animation-iteration-count: infinite;
+              }
+              .f1t9cprh {
+                animation-duration: 5s;
+              }
+          `);
     });
 
+    it('allows to define array as animationName', () => {
+      expect(
+        resolveStyleRules({
+          animationName: [
+            {
+              from: {
+                transform: 'rotate(0deg)',
+              },
+              to: {
+                transform: 'rotate(360deg)',
+              },
+            },
+            {
+              from: {
+                opacity: 0,
+              },
+              to: {
+                opacity: 1,
+              },
+            },
+          ],
+          animationIterationCount: 'infinite',
+          animationDuration: '5s',
+        }),
+      ).toMatchInlineSnapshot(`
+               @-webkit-keyframes f13owpa8 {
+                 from {
+                   -webkit-transform: rotate(0deg);
+                   -moz-transform: rotate(0deg);
+                   -ms-transform: rotate(0deg);
+                   transform: rotate(0deg);
+                 }
+                 to {
+                   -webkit-transform: rotate(360deg);
+                   -moz-transform: rotate(360deg);
+                   -ms-transform: rotate(360deg);
+                   transform: rotate(360deg);
+                 }
+               }
+               @keyframes f13owpa8 {
+                 from {
+                   -webkit-transform: rotate(0deg);
+                   -moz-transform: rotate(0deg);
+                   -ms-transform: rotate(0deg);
+                   transform: rotate(0deg);
+                 }
+                 to {
+                   -webkit-transform: rotate(360deg);
+                   -moz-transform: rotate(360deg);
+                   -ms-transform: rotate(360deg);
+                   transform: rotate(360deg);
+                 }
+               }
+               @-webkit-keyframes f1qa61cu {
+                 from {
+                   opacity: 0;
+                 }
+                 to {
+                   opacity: 1;
+                 }
+               }
+               @keyframes f1qa61cu {
+                 from {
+                   opacity: 0;
+                 }
+                 to {
+                   opacity: 1;
+                 }
+               }
+               @-webkit-keyframes rf13owpa8 {
+                 from {
+                   -webkit-transform: rotate(0deg);
+                   -moz-transform: rotate(0deg);
+                   -ms-transform: rotate(0deg);
+                   transform: rotate(0deg);
+                 }
+                 to {
+                   -webkit-transform: rotate(-360deg);
+                   -moz-transform: rotate(-360deg);
+                   -ms-transform: rotate(-360deg);
+                   transform: rotate(-360deg);
+                 }
+               }
+               @keyframes rf13owpa8 {
+                 from {
+                   -webkit-transform: rotate(0deg);
+                   -moz-transform: rotate(0deg);
+                   -ms-transform: rotate(0deg);
+                   transform: rotate(0deg);
+                 }
+                 to {
+                   -webkit-transform: rotate(-360deg);
+                   -moz-transform: rotate(-360deg);
+                   -ms-transform: rotate(-360deg);
+                   transform: rotate(-360deg);
+                 }
+               }
+               .f18gdskf {
+                 animation-name: f13owpa8 f1qa61cu;
+               }
+               .rf18gdskf {
+                 animation-name: rf13owpa8 f1qa61cu;
+               }
+               .f1cpbl36 {
+                 animation-iteration-count: infinite;
+               }
+               .f1t9cprh {
+                 animation-duration: 5s;
+               }
+          `);
+    });
+  });
+
+  describe('experimental', () => {
     it('allows to increase specificity', () => {
       expect(resolveStyleRules({ color: 'red' }, 1)).toMatchInlineSnapshot(`
         .fe3e8s901.fe3e8s901 {