+++ packages/react-avatar/src/components/Avatar/Avatar.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useAvatar } from './useAvatar';
 import { AvatarProps } from './Avatar.types';
 import { makeClasses } from '@fluentui/react-utils';
-import { useInlineTokens } from '@fluentui/react-theme-provider';
+import { useInlineTokens } from '@fluentui/react-theme-provider/lib/compat/index';
 import { useFocusRects, nullRender } from '@fluentui/utilities';
 import * as classes from './Avatar.scss';
 import { Badge } from '../Badge/Badge';