+++ packages/react-theme-provider/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-theme-provider
 
-This log was last generated on Mon, 01 Feb 2021 12:23:48 GMT and should not be manually modified.
+This log was last generated on Tue, 02 Feb 2021 12:21:54 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.0-beta.19](https://github.com/microsoft/fluentui/tree/@fluentui/react-theme-provider_v1.0.0-beta.19)
+
+Tue, 02 Feb 2021 12:21:54 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-theme-provider_v1.0.0-beta.18..@fluentui/react-theme-provider_v1.0.0-beta.19)
+
+### Changes
+
+- move react-theme-provider utils to /compat ([PR #16729](https://github.com/microsoft/fluentui/pull/16729) by olfedias@microsoft.com)
+- feat: add converged Provider and basic theme utils ([PR #16762](https://github.com/microsoft/fluentui/pull/16762) by olfedias@microsoft.com)
 
 ## [1.0.0-beta.18](https://github.com/microsoft/fluentui/tree/@fluentui/react-theme-provider_v1.0.0-beta.18)